Add shop list page
Added ShopList component and used it in the home page.
The ShopList component has ShopPreview and VPagination as child components.
The ShopPreview handle the display of each shop
and it has a child component too, the ShopActions component.
The ShopActions handle like/dislike action of  the shop.
Added unit test for the ShopList component.